fix(apollo-react): add stage execution states with adhoc play button
