+++ packages/apollo-react/src/canvas/components/StageNode/StageTaskDragOverlay.tsx
@@ -2,9 +2,9 @@ import { DragOverlay } from '@dnd-kit/core';
 import { useStore } from '@uipath/apollo-react/canvas/xyflow/react';
 import { useMemo } from 'react';
 import { createPortal } from 'react-dom';
-import { TaskContent } from './DraggableTask';
 import { StageTask } from './StageNode.styles';
 import type { StageTaskDragOverlayProps } from './StageNode.types';
+import { TaskContent } from './TaskContent';
 
 export const StageTaskDragOverlay = ({
   activeTask,
